requirements.txt
api==0.0.7
asgiref==3.7.1
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
Django==4.2.1
djangorestframework==3.14.0
idna==3.4
nose==1.3.7
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.6.1
urllib3==2.0.2
